requirements.txt
antiorm==1.2.1
db==0.1.1
db-sqlite3==0.0.1
numpy==2.0.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
pillow==10.4.0
pytest
pylint
black